Fix sisu data schema to use organization share as float

+++ services/headless-lms/utils/src/services/sisu.rs
@@ -160,7 +160,7 @@ pub struct Organisation {
     pub organisation_id: Option<String>,
     pub educational_institution_urn: Option<serde_json::Value>,
     pub role_urn: String,
-    pub share: i64,
+    pub share: f64,
     pub validity_period: Option<OrganisationValidityPeriod>,
 }
